requirements.txt
# ── Core ──────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# ── LangChain / LangGraph ─────────────────
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
langgraph==0.2.50
langfuse==2.50.0

# ── LLM local ─────────────────────────────
ollama==0.3.3

# ── Vector DB ─────────────────────────────
chromadb==0.5.20
sentence-transformers==3.3.1

# ── Visión ────────────────────────────────
opencv-python==4.10.0.84
Pillow==11.0.0
scikit-image==0.24.0
imageio==2.36.1
numpy==1.26.4

# ── ML ────────────────────────────────────
torch==2.4.1
torchvision==0.19.1

# ── YOLO / Segmentación ───────────────────
ultralytics==8.3.40
supervision==0.24.0

# ── Corpus / NLP ──────────────────────────
pymupdf==1.24.14
easyocr==1.7.2
openai-whisper==20240930
spacy==3.8.3

# ── API ───────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
httpx==0.27.2

# ── Bot ───────────────────────────────────
aiogram==3.14.0

# ── Documentación ─────────────────────────
jinja2==3.1.4
weasyprint==62.3

# ── Infra ─────────────────────────────────
redis==5.2.1
celery==5.4.0
minio==7.2.11
sqlalchemy==2.0.36
asyncpg==0.30.0

# ── Testing ───────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
